refactor: move collapseSystemInPlace to src/utils/
Colocates with other small focused utilities. Trims caller-specific
details from JSDoc (belongs at call site, not in the utility).
