+++ src/utils/system-collapse.test.ts
